Remove rng_config
